Besluitgebied_A
Besluitgebied_A is either a ro:AMVB or ro:Regeling.
It has a ro:Planobject; ro:Besluitvlak.
ro:Besluitvlak is constrained according to the respective plan type.

ro:Besluitvlak_A and Besluitgebied_A contain all the properties not applicable to each specialisation of ro:Besluit.

It does not use dc:subject instead of ro:thema.

ro:kaart is not (yet) used as it is not clear what its relation is with ro:Illustratie